Separate code into two components
